just download new Senate pdf register for now

+++ update.sh
@@ -19,7 +19,7 @@ urlsenat=$(curl http://www.senat.fr/role/groupes_interet.html 2> /dev/null |
   iconv -f iso-8859-15 -t utf-8 |
   grep "Liste des représentants d'" |
   head -n 1 |
-  sed 's|^.* href="\([^"]*xlsx\?\)".*$|http://www.senat.fr\1|')
+  sed 's#^.* href="\([^"]*\.\(xlsx\?\|pdf\)\)".*$#http://www.senat.fr\1#')
 filename=$(echo $urlsenat | sed 's|^.*/\([^/]\+\)$|data/\1|')
 if ! [ -z "$filename" ] && ! test -f "$filename"; then
   echo 
@@ -27,12 +27,20 @@ if ! [ -z "$filename" ] && ! test -f "$filename"; then
   echo "DOWNLOADING NEW SENATE REGISTER VERSION"
   echo "---------------------------------------"
   wget "$urlsenat" -O "$filename"
-  in2csv -H "$filename" > "$filename.csv"
-  ./clean_senat.py "$filename.csv" > data/registre-lobbying-Senat.csv
-  rm -f "$filename.csv"
-  if ! $DEBUG; then
-    git commit data/registre-lobbying-Senat.csv -m "update registre Sénat"
-    git push
+  if echo $urlsenat | grep xls > /dev/null; then
+    in2csv -H "$filename" > "$filename.csv"
+    ./clean_senat.py "$filename.csv" > data/registre-lobbying-Senat.csv
+    rm -f "$filename.csv"
+    if ! $DEBUG; then
+      git commit data/registre-lobbying-Senat.csv -m "update registre Sénat"
+      git push
+    fi
+  else
+    cp -f "$filename" "data/registre-lobbying-Senat.pdf"
+    if ! $DEBUG; then
+      git commit data/registre-lobbying-Senat.pdf -m "update registre Sénat"
+      git push
+    fi
   fi
 fi
